testmethods/requirements.txt
selenium
pytest
pytest-allure-adaptor
pytest-bdd